% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_labels.R
\docType{data}
\name{team_labels}
\alias{team_labels}
\title{Team labels dataset}
\format{
A character vector with one label per team.
}
\usage{
team_labels
}
\description{
A dataset containing labels for each MLB team with year ranges and
number of home stadiums. Generated internally by `get_team_labels()`.
}
\examples{
data(team_labels)
head(team_labels)
}
\keyword{datasets}
